+++ libs/base_core/src/index.rs
@@ -19,7 +19,7 @@
 #![cfg_attr(all(nightly_doc, miri), allow(unused_attributes))]
 #![cfg_attr(all(nightly_doc, not(doc)), allow(unused_attributes))]
 #![cfg_attr(nightly_float, feature(f16, f128))]
-// #![cfg_attr(nightly_simd, feature(portable_simd))]
+#![cfg_attr(nightly_simd, feature(portable_simd))]
 
 /* crate safeguards */