+++ libs/base_core/src/num/wide/mod.rs
@@ -5,10 +5,14 @@
 
 mod lane; // define_lane!
 
+mod reexports;
+
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             lane::*,
         };
+
+        pub use super::reexports::*;
     }
 }